+++ srcs/readline/get_cursor_pos_3.c
@@ -13,9 +13,9 @@
 #include "readline.h"
 #include "libft.h"
 
-int	get_cursor_pos_next_alnum(t_matrix *matrix)
+size_t	get_cursor_pos_next_alnum(t_matrix *matrix)
 {
-	int		col;
+	size_t	col;
 	t_line	*line;
 
 	col = matrix->cursor->col;
@@ -27,9 +27,9 @@ int	get_cursor_pos_next_alnum(t_matrix *matrix)
 	return (col);
 }
 
-int	get_cursor_pos_begin_alnum(t_matrix *matrix)
+size_t	get_cursor_pos_begin_alnum(t_matrix *matrix)
 {
-	int		col;
+	size_t	col;
 	t_line	*line;
 
 	col = matrix->cursor->col;
@@ -45,9 +45,9 @@ int	get_cursor_pos_begin_alnum(t_matrix *matrix)
 	return (col);
 }
 
-int	get_cursor_pos_end_alnum(t_matrix *matrix)
+size_t	get_cursor_pos_end_alnum(t_matrix *matrix)
 {
-	int		col;
+	size_t	col;
 	t_line	*line;
 
 	col = matrix->cursor->col;
