+++ srcs/readline/line_resize.c
@@ -13,7 +13,7 @@
 #include "readline.h"
 #include "shell.h"
 
-void	line_resize(t_line *line, int new_size, int old_size)
+void	line_resize(t_line *line, size_t new_size, size_t old_size)
 {
 	line->buf = (char *)xrealloc(line->buf, new_size + 1, old_size);
 	line->size = new_size;
