+++ srcs/readline/matrix_resize.c
@@ -13,7 +13,7 @@
 #include "readline.h"
 #include "shell.h"
 
-void	matrix_resize(t_matrix *matrix, int new_size, int old_size)
+void	matrix_resize(t_matrix *matrix, size_t new_size, size_t old_size)
 {
 	matrix->lines = (t_line **)xrealloc(matrix->lines,
 			sizeof(t_line *) * new_size, sizeof(t_line *) * old_size);
